+++ elliott/elliottlib/cli/__main__.py
@@ -68,7 +68,7 @@ from elliottlib.cli.get_golang_report_cli import get_golang_report_cli
 from elliottlib.cli.snapshot_cli import snapshot_cli
 from elliottlib.cli.konflux_release_cli import konflux_release_cli
 from elliottlib.cli import konflux_release_watch_cli
-from elliottlib.cli.pin_cli import assembly_pin_builds_cli
+from elliottlib.cli.pin_builds_cli import assembly_pin_builds_cli
 
 # 3rd party
 import click
